Added some tools and scripts
Added basic scripts and functions in python that were used to streamline the process. Added directories of some of the BioInfo tools used (Smart PCA, Metal Meta-Analysis, GenotypeHarmonizer used for alignment, Plink executables for Linux, GCTA tool used in conditional analysis and heritability etc). The expression data files for eQTL analysis are very large, and therefore couldn't be uploaded as is, but an image showing the files used and the structure has been added. The files themselves can be shared over SSH etc as and when required by some collaborator or researcher or publisher.